Dao Finish

+++ src/main/java/model/Endereco.java
@@ -2,7 +2,7 @@ package model;
 
 public class Endereco {
 
-    private long id;
+    private int id;
     private String rua;
     private String numero;
     private String complemento;
@@ -13,11 +13,11 @@ public class Endereco {
     public Endereco() {
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
 
@@ -67,5 +67,5 @@ public class Endereco {
 
     public void setEstado(String estado) {
         this.estado = estado;
-    }
+    }    
 }
